+++ tests/ngr/elixir/test/test_helper.exs
@@ -1,0 +1,2 @@
+# https://hexdocs.pm/ex_unit/ExUnit.html#module-integration-with-mix
+ExUnit.start()