+++ CHANGES.md
@@ -6,6 +6,7 @@
 - ngr: Improve runner for Python projects
   Also consider setup.py, setup.cfg, pyproject.toml, and poetry.lock.
 - ngr: Add runner for Golang projects
+- ngr: Add runner for Elixir projects
 
 ## 2023-10-30 v0.0.1
 - Add `pueblo.util.logging.setup_logging`, nomen est omen